Abstract responsive props as a Vue mixin.
